Return calculated speed from single testers, remove hostnames from error msgs

+++ src/services/speedtest.h
@@ -23,8 +23,8 @@ void fill_buffer(UploadData *buf);
 
 char *probe_download_endpoint(CURL *curl, const char *host, size_t timeout, int disableSSL);
 
-int get_download_speed_single(CURL *curl, Server *server, size_t timeout, int disableSSL);
-int get_upload_speed_single(CURL *curl, Server *server, size_t timeout, int disableSSL);
+double get_download_speed_single(CURL *curl, Server *server, size_t timeout, int disableSSL);
+double get_upload_speed_single(CURL *curl, Server *server, size_t timeout, int disableSSL);
 int get_download_speed(Server *servers, size_t count, size_t timeout, int disableSSL);
 int get_upload_speed(Server *servers, size_t count, size_t timeout, int disableSSL);
